move completions out to own module
